Including dijkstra_arithmetics and stack_with_max
Including dijkstra algorithm for evaluating arithmetic simple
expressions as suggested in the Algorithm course. Also, implementing the
stack variant that allows for efficient max() as part of one of the
exercises. The implementation can be improved a lot with a good peek()
method that I have not implemented yet for the basic Stack