Including Quick Sort and Selection
Quick Sort implementation included.
Changing implementations from other algorithms to just work with
slices also as it is more general.
For Quick Sort, an algorithm without high and low params is used.
However, for selection, I implement the algorithm with high and lo
because it needs to keep comparing with k and, for that purpose,
sub-slices don't work well